bert toy example and fixes
